+++ src/library/DIPS.Mobile.UI/Effects/Layout/Android/LayoutPlatformEffect.cs
@@ -1,6 +1,7 @@
 using Android.Graphics.Drawables;
 using Google.Android.Material.Shape;
 using Microsoft.Maui.Platform;
+using CollectionView = DIPS.Mobile.UI.Components.Lists.CollectionView;
 using Colors = Microsoft.Maui.Graphics.Colors;
 
 namespace DIPS.Mobile.UI.Effects.Layout;
@@ -8,24 +9,21 @@ namespace DIPS.Mobile.UI.Effects.Layout;
 public partial class  LayoutPlatformEffect
 {
     private Drawable? m_originalBackground;
-    
-    protected override partial void OnAttached()
+
+    private partial void PlatformOnAttached(CornerRadius cornerRadius)
     {
         m_originalBackground = Control.Background;
-        
-        var shapeAppearanceModel = new ShapeAppearanceModel.Builder().SetAllCorners(CornerFamily.Rounded, (float)(Layout.GetCornerRadius(Element).TopLeft * Control.Context.GetDisplayDensity())).Build();
-        var materialShapeDrawable = new MaterialShapeDrawable(shapeAppearanceModel);
+
+        var materialShapeDrawable = MaterialShapeDrawableHelper.GetMaterialShapeDrawableFromCornerRadius(cornerRadius);
       
         Control.ClipToOutline = true;
-
+        
         if (Element is VisualElement visualElement)
         {
             materialShapeDrawable.FillColor = visualElement.BackgroundColor is not null ? 
                 visualElement.BackgroundColor.ToDefaultColorStateList() 
                 : Colors.Transparent.ToDefaultColorStateList();
         }
-        
-        materialShapeDrawable.StrokeWidth = 0;
         
         Control.Background = materialShapeDrawable;
     }